Pass proper context to renderers
