+++ Contentful.podspec
@@ -29,14 +29,18 @@ Pod::Spec.new do |spec|
   spec.tvos.source_files         = 'Sources/Contentful/UIKit/*.swift'
   spec.osx.source_files          = 'Sources/Contentful/AppKit/*.swift'
  
-  spec.ios.deployment_target     = '8.0'
+  spec.ios.deployment_target     = '9.3'
   spec.osx.deployment_target     = '10.10'
   spec.watchos.deployment_target = '2.0'
   spec.tvos.deployment_target    = '9.0'
 
   # Subspecs
   spec.subspec 'ImageOptions' do |subspec|
-    subspec.source_files = 'Sources/Contentful/ImageOptions.swift'
+    subspec.source_files = 'Sources/Contentful/ImageOptions/*.swift'
+  end
+
+  spec.subspec 'StructuredTextKit' do |subspec|
+    subspec.source_files = 'Sources/Contentful/StructuredTextKit/*.swift'
   end
 end
