+++ Sources/Contentful/ImageOptions/ImageOptions.swift
